worker: Extract 'job_registry`, `util` and `worker` modules
