+++ crates_io_worker/src/lib.rs
@@ -1,8 +1,11 @@
 mod background_job;
 mod errors;
+mod job_registry;
 mod runner;
 pub mod schema;
 mod storage;
+mod util;
+mod worker;
 
 pub use self::background_job::BackgroundJob;
 pub use self::errors::EnqueueError;
