+++ crates_io_worker/src/util.rs
@@ -1,0 +1,39 @@
+use anyhow::anyhow;
+use std::any::Any;
+use std::panic::PanicInfo;
+
+pub fn with_sentry_transaction<F, R, E>(transaction_name: &str, callback: F) -> Result<R, E>
+where
+    F: FnOnce() -> Result<R, E>,
+{
+    let tx_ctx = sentry_core::TransactionContext::new(transaction_name, "swirl.perform");
+    let tx = sentry_core::start_transaction(tx_ctx);
+
+    let result = sentry_core::with_scope(|scope| scope.set_span(Some(tx.clone().into())), callback);
+
+    tx.set_status(match result.is_ok() {
+        true => sentry_core::protocol::SpanStatus::Ok,
+        false => sentry_core::protocol::SpanStatus::UnknownError,
+    });
+    tx.finish();
+
+    result
+}
+
+/// Try to figure out what's in the box, and print it if we can.
+///
+/// The actual error type we will get from `panic::catch_unwind` is really poorly documented.
+/// However, the `panic::set_hook` functions deal with a `PanicInfo` type, and its payload is
+/// documented as "commonly but not always `&'static str` or `String`". So we can try all of those,
+/// and give up if we didn't get one of those three types.
+pub fn try_to_extract_panic_info(info: &(dyn Any + Send + 'static)) -> anyhow::Error {
+    if let Some(x) = info.downcast_ref::<PanicInfo<'_>>() {
+        anyhow!("job panicked: {x}")
+    } else if let Some(x) = info.downcast_ref::<&'static str>() {
+        anyhow!("job panicked: {x}")
+    } else if let Some(x) = info.downcast_ref::<String>() {
+        anyhow!("job panicked: {x}")
+    } else {
+        anyhow!("job panicked")
+    }
+}
